+++ tests/testthat/test-models.R
@@ -1,0 +1,4 @@
+test_that("fw_model() works", {
+    expect_identical(fw_model(), mod_lv_fr1)
+})
+