+++ NEWS.md
@@ -1,5 +1,7 @@
 # fwebinfr devel 
 
+* Add `fw_create()` to create objects of class `fwmod`.
+* Add `fw_model()` to select model (only one currently available).
 * Export `fw_get_A_predicted()` and `fw_get_B_predicted()`.
 * Add vignette "Get started with fwebinfr".
 * `fw_infer()` gains arguments `sdB` to dealth with biomass variation and `...`